upload new lesson
-   add package react-router@7
-   update main.jsx : add react-router BrowserRouter
-   create folder routers
-   create folder routers/home
-   create file to routers/home add home.component.jsx
-   update App.jsx : copy all code
    -   to directory.component.jsx
-   update directory.component.jsx : App String change to Home
-   update App.jsx : import Home
-   update App.jsx : add react-router Route & Routes
-   update App.jsx : set Home Route
-   update vite.config.js : add rouers link
-   update jsconfig.json : add rouers link
