+++ jsconfig.json
@@ -4,6 +4,7 @@
     "paths": {
       "@/*": ["src/*"],
       "@components/*": ["src/components/*"],
+      "@routers/*": ["src/routers/*"],
     }
   },
   "exclude": ["node_modules", "dist"]
