+++ package.json
@@ -13,6 +13,7 @@
 	"dependencies": {
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
+		"react-router": "^7.7.1",
 		"sass": "^1.89.2"
 	},
 	"devDependencies": {
