+++ src/App.jsx
@@ -1,34 +1,16 @@
-import Directory from '@components/directory/directory.component';
+import { Routes, Route } from 'react-router';
+
+import Home from '@routers/home/home.component';
 
 function App() {
-	const categories = [
-		{
-			id: 1,
-			title: 'hats',
-			imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
-		},
-		{
-			id: 2,
-			title: 'jackets',
-			imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
-		},
-		{
-			id: 3,
-			title: 'sneakers',
-			imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
-		},
-		{
-			id: 4,
-			title: 'womens',
-			imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
-		},
-		{
-			id: 5,
-			title: 'mens',
-			imageUrl: 'https://i.ibb.co/R70vBrQ/men.png',
-		},
-	];
-	return <Directory categories={categories} />;
+	return (
+		<Routes>
+			<Route
+				path="/"
+				element={<Home />}
+			/>
+		</Routes>
+	);
 }
 
 export default App;
